Rename and move validation schemas
