+++ lib/applicants/validation.ts
@@ -1,10 +1,10 @@
 import { ApplicationType, Gender } from '@prisma/client';
-import { boolean, date, mixed, object, string } from 'yup';
+import { boolean, date, mixed, number, object, string } from 'yup';
 
 /**
  * Validation schema for permit holder information forms when creating and viewing requests
  */
-export const requestingPermitHolderInformationSchema = object({
+export const requestPermitHolderInformationSchema = object({
   permitHolder: object({
     type: mixed<ApplicationType>().oneOf(Object.values(ApplicationType)).required(),
     firstName: string().required('Please enter a first name'),
@@ -67,4 +67,20 @@ export const permitHolderInformationSchema = object({
  */
 export const newPermitHolderInformationSchema = object({
   permitHolder: permitHolderInformationSchema,
+});
+
+/**
+ * Verify identity validation schema
+ */
+export const verifyIdentitySchema = object({
+  userId: number()
+    .typeError('User ID must be a positive number')
+    .required('Please enter your User ID')
+    .positive('Please enter your User ID')
+    .integer('Please enter your User ID'),
+  phoneNumberSuffix: string()
+    .matches(/^\d+$/, 'Must only contain numbers')
+    .required('Please enter the last 4 digits of your phone number')
+    .length(4, 'Must be exactly 4 digits'),
+  dateOfBirth: date().required('Please enter your date of birth'),
 });
